+++ duckscript_sdk/src/sdk/std/string/mod.rs
@@ -1,7 +1,9 @@
 mod contains;
 mod ends_with;
 mod equals;
+mod is_empty;
 mod starts_with;
+mod trim;
 
 use crate::utils::pckg;
 use duckscript::types::command::Commands;
@@ -15,7 +17,9 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(contains::create(&package))?;
     commands.set(ends_with::create(&package))?;
     commands.set(equals::create(&package))?;
+    commands.set(is_empty::create(&package))?;
     commands.set(starts_with::create(&package))?;
+    commands.set(trim::create(&package))?;
 
     Ok(())
 }